# public/swagger.json
{"openapi":"3.0.0","paths":{"/alunos":{"get":{"operationId":"AlunoController_findAll","parameters":[],"responses":{"200":{"description":"Lista de todos os alunos"}},"tags":["Alunos"]},"post":{"operationId":"AlunoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Aluno"}}}},"responses":{"201":{"description":"Aluno criado com sucesso"}},"tags":["Alunos"]}},"/alunos/{id}":{"get":{"operationId":"AlunoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Detalhes de um aluno"}},"tags":["Alunos"]},"put":{"operationId":"AlunoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Aluno"}}}},"responses":{"200":{"description":"Aluno atualizado com sucesso"}},"tags":["Alunos"]},"delete":{"operationId":"AlunoController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"Aluno excluído com sucesso"}},"tags":["Alunos"]}}},"info":{"title":"API de Alunos","description":"API para gerenciamento de alunos","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"Aluno":{"type":"object","properties":{"id":{"type":"number","description":"ID único do aluno.","example":1},"nome":{"type":"string","description":"Nome do aluno.","example":"João Silva"},"idade":{"type":"number","description":"Idade do aluno.","example":25},"notaPrimeiroSemestre":{"type":"number","description":"Nota do aluno no primeiro semestre.","example":8.5},"notaSegundoSemestre":{"type":"number","description":"Nota do aluno no segundo semestre.","example":9},"nomeProfessor":{"type":"string","description":"Nome do professor responsável pelo aluno.","example":"Prof. Maria Oliveira"},"numeroSala":{"type":"number","description":"Número da sala em que o aluno está matriculado.","example":102}},"required":["id","nome","idade","notaPrimeiroSemestre","notaSegundoSemestre","nomeProfessor","numeroSala"]}}}}